sites: add api endpoint for approving site join request
